pyproject.toml
[tool.poetry]
name = "perdix-aadhaar-service"
version = "0.1.0"
description = ""
authors = ["stalin.solomon <stalin.solomon@dvara.com>"]

[tool.poetry.dependencies]
python = "^3.10"
yattag = "^1.14.0"
peewee = "^3.14.8"
PyMySQL = "^1.0.2"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
pydantic = "^1.8.2"
pycryptodome = "^3.12.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
